I've tried and tried and tried to get CreateJS working so that I can preload sounds, but it just doesn't seem to work.
In assets.ts I've created a `loadSoundsAsync` function that uses `createjs.LoadQueue` to try and load a list of sound file according to the `sounds/sounds.json` files. I set up an event handler on the "complete" event but it is never triggered. Neither is "error". I can't figure out how. Sound I've not idea if sounds are being preloaded correctly or when they have completed.